Fixed tests

+++ WorldServer/World/Battlefronts/Apocalypse/Loot/IRewardAssigner.cs
@@ -4,6 +4,6 @@ namespace WorldServer.World.Battlefronts.Apocalypse.Loot
 {
     public interface IRewardAssigner
     {
-        List<LootBagTypeDefinition> AssignLootToPlayers(List<uint> eligiblePlayers, byte forceNumberRewards=0);
+        List<LootBagTypeDefinition> AssignLootToPlayers(List<uint> eligiblePlayers, int forceNumberRewards=0);
     }
 }
